feat: capability to remove unused elements of the file
