+++ internal/suggestions/example_experiment.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	b64 "encoding/base64"
 	"fmt"
+	"github.com/speakeasy-api/speakeasy/internal/transform"
 	"math/rand"
 	"net/http"
 	"os"
@@ -15,7 +16,6 @@ import (
 	"github.com/pb33f/libopenapi"
 	"github.com/pb33f/libopenapi/datamodel"
 	v3 "github.com/pb33f/libopenapi/datamodel/high/v3"
-	"github.com/pb33f/libopenapi/index"
 	"github.com/pb33f/libopenapi/orderedmap"
 	"github.com/speakeasy-api/openapi-generation/v2/pkg/errors"
 	"github.com/speakeasy-api/openapi-overlay/pkg/loader"
@@ -317,7 +317,7 @@ func Split(doc libopenapi.Document, cacheFolder string) ([]Shard, error) {
 		v3Model.Model.Paths.PathItems = orderedmap.New[string, *v3.PathItem]()
 		v3Model.Model.Paths.PathItems.Set(pair.Key(), pair.Value())
 		// eliminate all the now-orphaned schemas
-		doc, v3Model, err = removeOrphans(doc, v3Model)
+		doc, v3Model, err = transform.RemoveOrphans(doc, v3Model)
 		if err != nil {
 			return nil, err
 		}
@@ -337,196 +337,5 @@ func base64(key string) string {
 	return b64.StdEncoding.EncodeToString([]byte(key))
 }
 
-func removeOrphans(doc libopenapi.Document, model *libopenapi.DocumentModel[v3.Document]) (libopenapi.Document, *libopenapi.DocumentModel[v3.Document], error) {
-	_, doc, model, errs := doc.RenderAndReload()
-	// remove nil errs
-	var nonNilErrs []error
-	for _, e := range errs {
-		if e != nil {
-			nonNilErrs = append(nonNilErrs, e)
-		}
-	}
-	if len(nonNilErrs) > 0 {
-		return nil, nil, fmt.Errorf("failed to render and reload document: %v", errs)
-	}
 
-	components := model.Model.Components
-	context := model
-	allRefs := context.Index.GetAllReferences()
-	schemasIdx := context.Index.GetAllComponentSchemas()
-	responsesIdx := context.Index.GetAllResponses()
-	parametersIdx := context.Index.GetAllParameters()
-	examplesIdx := context.Index.GetAllExamples()
-	requestBodiesIdx := context.Index.GetAllRequestBodies()
-	headersIdx := context.Index.GetAllHeaders()
-	securitySchemesIdx := context.Index.GetAllSecuritySchemes()
-	linksIdx := context.Index.GetAllLinks()
-	callbacksIdx := context.Index.GetAllCallbacks()
-	mappedRefs := context.Index.GetMappedReferences()
 
-	checkOpenAPISecurity := func(key string) bool {
-		if strings.Contains(key, "securitySchemes") {
-			segs := strings.Split(key, "/")
-			def := segs[len(segs)-1]
-			for r := range context.Index.GetSecurityRequirementReferences() {
-				if r == def {
-					return true
-				}
-			}
-		}
-		return false
-	}
-
-	// create poly maps.
-	oneOfRefs := make(map[string]*index.Reference)
-	allOfRefs := make(map[string]*index.Reference)
-	anyOfRefs := make(map[string]*index.Reference)
-
-	// include all polymorphic references.
-	for _, ref := range context.Index.GetPolyAllOfReferences() {
-		allOfRefs[ref.Definition] = ref
-	}
-	for _, ref := range context.Index.GetPolyOneOfReferences() {
-		oneOfRefs[ref.Definition] = ref
-	}
-	for _, ref := range context.Index.GetPolyAnyOfReferences() {
-		anyOfRefs[ref.Definition] = ref
-	}
-
-	notUsed := make(map[string]*index.Reference)
-	mapsToSearch := []map[string]*index.Reference{
-		schemasIdx,
-		responsesIdx,
-		parametersIdx,
-		examplesIdx,
-		requestBodiesIdx,
-		headersIdx,
-		securitySchemesIdx,
-		linksIdx,
-		callbacksIdx,
-	}
-
-	for _, resultMap := range mapsToSearch {
-		for key, ref := range resultMap {
-
-			u := strings.Split(key, "#/")
-			keyAlt := key
-			if len(u) == 2 {
-				if u[0] == "" {
-					keyAlt = fmt.Sprintf("%s#/%s", context.Index.GetSpecAbsolutePath(), u[1])
-				}
-			}
-
-			if allRefs[key] == nil && allRefs[keyAlt] == nil {
-				found := false
-
-				if oneOfRefs[key] != nil || allOfRefs[key] != nil || anyOfRefs[key] != nil {
-					found = true
-				}
-
-				if mappedRefs[key] != nil || mappedRefs[keyAlt] != nil {
-					found = true
-				}
-
-				if !found {
-					found = checkOpenAPISecurity(key)
-				}
-
-				if !found {
-					notUsed[key] = ref
-				}
-			}
-		}
-	}
-
-	// let's start killing orphans
-	anyRemoved := false
-	schemas := components.Schemas
-	toDelete := make([]string, 0)
-	for pair := orderedmap.First(schemas); pair != nil; pair = pair.Next() {
-		// remove all schemas that are not referenced
-		if !isReferenced(pair.Key(), "schemas", notUsed) {
-			toDelete = append(toDelete, pair.Key())
-			fmt.Printf("dropped #/components/schemas/%s\n", pair.Key())
-			anyRemoved = true
-		}
-	}
-	for _, key := range toDelete {
-		schemas.Delete(key)
-	}
-
-	responses := components.Responses
-	toDelete = make([]string, 0)
-	for pair := orderedmap.First(responses); pair != nil; pair = pair.Next() {
-		// remove all responses that are not referenced
-		if !isReferenced(pair.Key(), "responses", notUsed) {
-			toDelete = append(toDelete, pair.Key())
-			fmt.Printf("dropped #/components/responses/%s\n", pair.Key())
-			anyRemoved = true
-		}
-	}
-	for _, key := range toDelete {
-		responses.Delete(key)
-	}
-	parameters := components.Parameters
-	toDelete = make([]string, 0)
-	for pair := orderedmap.First(parameters); pair != nil; pair = pair.Next() {
-		// remove all parameters that are not referenced
-		if !isReferenced(pair.Key(), "parameters", notUsed) {
-			toDelete = append(toDelete, pair.Key())
-			anyRemoved = true
-		}
-	}
-	for _, key := range toDelete {
-		parameters.Delete(key)
-	}
-	examples := components.Examples
-	toDelete = make([]string, 0)
-	for pair := orderedmap.First(examples); pair != nil; pair = pair.Next() {
-		// remove all examples that are not referenced
-		if !isReferenced(pair.Key(), "examples", notUsed) {
-			toDelete = append(toDelete, pair.Key())
-			anyRemoved = true
-		}
-	}
-	for _, key := range toDelete {
-		examples.Delete(key)
-	}
-	requestBodies := components.RequestBodies
-	toDelete = make([]string, 0)
-	for pair := orderedmap.First(requestBodies); pair != nil; pair = pair.Next() {
-		// remove all requestBodies that are not referenced
-		if !isReferenced(pair.Key(), "requestBodies", notUsed) {
-			toDelete = append(toDelete, pair.Key())
-			anyRemoved = true
-		}
-	}
-	for _, key := range toDelete {
-		requestBodies.Delete(key)
-	}
-
-	headers := components.Headers
-	toDelete = make([]string, 0)
-	for pair := orderedmap.First(headers); pair != nil; pair = pair.Next() {
-		// remove all headers that are not referenced
-		if !isReferenced(pair.Key(), "headers", notUsed) {
-			toDelete = append(toDelete, pair.Key())
-			anyRemoved = true
-		}
-	}
-	for _, key := range toDelete {
-		headers.Delete(key)
-	}
-	if anyRemoved {
-		return removeOrphans(doc, model)
-	}
-	return doc, model, nil
-}
-
-func isReferenced(key string, within string, notUsed map[string]*index.Reference) bool {
-	ref := fmt.Sprintf("#/components/%s/%s", within, key)
-	if notUsed[ref] != nil {
-		return false
-	}
-	return true
-}
